backend/pyproject.toml
[tool.ruff]
line-length = 120
ignore = ["I001"]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    #PEP8-Naming
    "N",
]

[tool.poetry]
name = "openeu-backend"
version = "0.0.1"
description = ""
authors = ["CSEE"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.13"
pre-commit = "3.7.1"
python-dotenv = "1.0.1"
fastapi = "0.115.12"
poetry-core = "2.1.3"
supabase = "2.15.1"
mypy = "1.15.0"
requests = "2.32.3"
types-requests = "2.32.0.20250328"
bs4 = "0.0.2"
scrapy = "2.13.0"
parsel = "1.10.0"
uvicorn = "^0.34.2"
crawl4ai = "0.6.3"
types-python-dateutil = "2.9.0.20250516"
deepl = "^1.22.0"
brevo-python = "^1.1.2"
rapidfuzz = "^3.13.0"
tiktoken = "^0.9.0"
openai = "^1.82.0"
scrapy-playwright = "^0.0.43"
playwright = "^1.52.0"
pygit2 = "1.18.0"
keybert = "^0.9.0"
scikit-learn = "^1.6.1"
sentence-transformers = "^4.1.0"
jinja2 = "3.1.3"
langchain = "^0.3.25"
schedule = "^1.2.2"
python-jose = "3.5.0"
langdetect = "^1.0.9"
psycopg2-binary = "^2.9.10"
fastapi-cache2 = "0.2.2"
cohere = "^5.15.0"
pandas = "^2.3.0"
openpyxl = "^3.1.5"
langchain-community = "^0.3.27"
pypdf = "^5.7.0"
python-docx = "^1.2.0"
email-validator = "^2.3.0"

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"
